mononoke: save latest synced changesets in blobstore instead of db
Summary:
In order to sync changesets we need to know what was the latest synced
changeset from a given source. Previously we wanted to store this data in sql,
however later we decided to store it in the file in the repo. The main reason
for doing that is that storing data inside the repo gives us both blame and
history, while database can't provide that.

This diff implements that - removes all the sql-related code for storing latest
synced changesets and instead puts the data inside the file in the repository.
Note that we still using sql to store mapping between (source, target)
changesets.

Reviewed By: mitrandir77

Differential Revision: D28470102

fbshipit-source-id: 996905b17ad4b0d5b0ea1e40c73d762850e113a8