mononoke/warm_bookmarks_cache: use a BookmarksSubscription to `init_bookmarks`
Summary:
Like it says in the title, this starts the work of incorporating
BookmarksSubscription into the WarmBookmarksCache. To do so, we start by using
the subscription to initialize bookmarks. Under the hood, this makes the same
SQL query that initializing the cache currently it does (except it does it in a
transaction to also get the corresponding log id).

Reviewed By: StanislavGlebik

Differential Revision: D28505643

fbshipit-source-id: 65c25acfd91bf36809602d75c55e5f74c1c54015