blame_v2: add "skip past this change" metadata
Summary:
Extend the `blame_v2` format to include metadata about the location in the
parent commit that a blamed line replaces.  This can be used to implement
accurate "skip past this change" in clients.

Most ranges only need the range of lines that the original blame range
replaces.  For ranges that are inserts, the parent range is of zero length and
the offset indicates the line that the range was inserted before.

For renames, we must include the path of the file before the rename, so that
the file can be found in the parent.

For merge commits, if the file is present in more than one parent, then lines
that are introduced in the merge commit itself have multiple possibilities for
the parent range.  We select and record the first parent that contains the file
as the provider of the parent range for these lines.  This favours the p1
history of the file, but allows "skip past this change" to work when files
are merged in.

Reviewed By: farnz

Differential Revision: D28546768

fbshipit-source-id: 2af1e95a0d27fb25aeea51682177fbac2c41b029