mononoke/wireproto_sink: track last failed io not just last failed flush
Summary:
In practice, if the client has disconnected it's unlikely that the thing that
will fail is a flush (instead if will probably be the poll that checks if we
can write before). So, let's track that separately.

This isn't super important since we can already infer the timing from when the
error was logged but it'll make the logs less ambiguous.

Reviewed By: johansglock

Differential Revision: D28637258

fbshipit-source-id: 3bc9c9aaa9fc8cf7a2d2514fb520cb1729f4c560