mononoke: implement add_sync_target
Summary:
This is the very first version of add_sync_target method. It's very naive - it
just creates a single merge commit, and it doesn't handle all the necessary
functionality (e.g. linkfiles) and it doesn't do all the necessary checks yet.
TODOs show the missing functionality.

However this implementation it should be good enough as the first version, and
we can continue to extend it.

Reviewed By: mitrandir77

Differential Revision: D28677365

fbshipit-source-id: a8af6b92ced285ce4d012843f591d5c29584669d