integration: run the hg integration tests with NFS too
Summary:
Similarly to the work that kmancini did for the integration tests, this
enables the Mercurial integration tests to run against NFS. Only 3 tests are
failing due to 3 separate reasons: graceful restart isn't support, permissions
given to the clients are wrong, and a lifetime issue for the Nfsd3 object.

As a bonus, I've cleaned the code a bit since TreeOnly was the only supported
configuration (and will likely stay that way).

Reviewed By: chadaustin

Differential Revision: D28459245

fbshipit-source-id: 21f043bb054d684eacaee57179f0bdf4853b0a35