add new types for storing params and responses
Summary:
Instead of storing params separate types for each type of request let's store
them in a single union type. This is going to simplify things and improve the
type safety (as we will no longer need to trust in the "request_type" string
store in sql to safely deserialize the type).

Reviewed By: krallin

Differential Revision: D28606341

fbshipit-source-id: cfa8ee15213e6582891906e326d0e258c3a918a1