mononoke: test and fix for hg deferred in OldestFirst direction
Summary:
When walking in OldestFirst direction, if we restart from a checkpoint there are some nodes that will never be part of a chunk visited by the current run. e.g. if we start at commit E and walk filenode that points to commit D it will be deferred, but as we are walking from old towards new, commit D will never be visited.

Fix is to not record deferred edges when walking in OldestFirst direction.

Reviewed By: farnz

Differential Revision: D28714771

fbshipit-source-id: 11faa2be62e588a284045bbdfffeb8f2a4efdf70