mononoke: don't record Changeset and ChangesetInfo deferred for Direction::NewestFirst
Summary:
When walking in OldestFirst direction, if we restart from a checkpoint there are some nodes that will never be part of a chunk visited by the current run. e.g. if we start at commit E and parent points to commit D it will be deferred, but as we are walking from old towards new, commit D will never be visited.

Fix is to not record deferred edges when walking in OldestFirst direction.

Previous similar change did this for unodes and filenodes, but also needed here.

Differential Revision: D28749514

fbshipit-source-id: e14a67cad38dfebba7a0b4cf9ffa3de63d8c41ce