glob benchmark: make the watchman and eden more comparable
Summary:
In the eden benchmark, EdenFS didn't return the file list, while Watchman would
always do it, thus let's always return the file list.

I've also removed the calls to SetItemsProcessed as the output is completely
misleading. The "Time" column is what needs to be looked at and not the
"items_per_second" that SetItemsProcessed shows.

Reviewed By: kmancini

Differential Revision: D28664475

fbshipit-source-id: e920b311464667a720d9ab00286e094829e292f7