mononoke: add InlineBase64 chunking to sqlblob
Summary:
Storing small blobs as separate chunks has the overhead of looking up chunk shard, wait for shard replication lag, chunk insert and/or generation update.

This adds a new chunking method that can store many small blobs inline in the the data table, as we have VARCHAR(255) available.  It Base64 encodes so as not to require changing schemas to a binary type.

Base64 fit 6 bits per byte so max stored in the VARCHAR column will be 255 * 6/8 = 191 bytes.   This is plenty to hold an empty packblob or a hash reference (e.g an alias git sha1 is 37 bytes )

Differential Revision: D28756290

fbshipit-source-id: 1ae545f4e333d92b9818c4170e21cd539ca66f94